Alle model types implementiert zur verwendung in anderen Klassen
